Added NuGetScrapingClient implementation to get download numbers from the package details page (for more up to date numbers). Changed AddHttpClient to use the scraping client rather than search api client.

+++ src/D20Tek.NuGet.Portfolio/Abstractions/INuGetSearchClient.cs
@@ -2,5 +2,5 @@
 
 public interface INuGetSearchClient
 {
-    Task<Result<int>> GetTotalDownloadsAsync(string packageId);
+    Task<Result<long>> GetTotalDownloadsAsync(string packageId);
 }
